chore: move cee
