+++ api/src/pdc/services/cee/contracts/common/CeeApplicationInterface.ts
